Avoid the generic promisify.
The doxbee and parallel benchmarks were using a generic promisify
implementation to wrap dummy functions with callbacks into promisified
functions that can be used with Promise/async code. This promisify
implementation doesn't match the implementation in Node.js, which is
not portable across engines (and somewhat coupled to Node.js internals).

However with the recent performance improvements the doxbee benchmark
is not dominated by this generic promisify, which means it doesn't
really measure Promise performance improvements anymore. So to make
sure the benchmarks are still meaningful we change them and manually
promisfy the dummy functions instead.

This will create a new baseline for the performance, but it'll make
sure that the benchmarks stay relevant.